Move lib fns to lib.rs
